temp fixes before step two of the integration
