+++ ha-core-infra/src/main/java/org/kie/remote/impl/consumer/InfraFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.hacep.core;
+package org.kie.remote;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -21,28 +21,17 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 
-import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.common.PartitionInfo;
 import org.apache.kafka.common.TopicPartition;
-import org.kie.hacep.EnvConfig;
-import org.kie.hacep.consumer.DroolsConsumerHandler;
-import org.kie.hacep.core.infra.DefaultSessionSnapShooter;
-import org.kie.hacep.core.infra.SessionSnapshooter;
-import org.kie.hacep.core.infra.consumer.ConsumerHandler;
-import org.kie.hacep.core.infra.consumer.DefaultKafkaConsumer;
-import org.kie.hacep.core.infra.consumer.EventConsumer;
-import org.kie.hacep.core.infra.consumer.ItemToProcess;
-import org.kie.hacep.core.infra.consumer.LocalConsumer;
-import org.kie.hacep.util.ConsumerUtilsCoreImpl;
-import org.kie.remote.RemoteKieSession;
-import org.kie.remote.RemoteStreamingKieSession;
-import org.kie.remote.TopicsConfig;
+
 import org.kie.remote.impl.ClientUtils;
 import org.kie.remote.impl.RemoteKieSessionImpl;
 import org.kie.remote.impl.RemoteStreamingKieSessionImpl;
-
+import org.kie.remote.impl.consumer.KafkaListenerThread;
 import org.kie.remote.impl.consumer.Listener;
+import org.kie.remote.impl.consumer.ListenerThread;
+import org.kie.remote.impl.consumer.LocalListenerThread;
 import org.kie.remote.impl.producer.EventProducer;
 import org.kie.remote.impl.producer.LocalProducer;
 import org.kie.remote.impl.producer.Producer;
@@ -50,22 +39,9 @@ import org.kie.remote.impl.producer.Producer;
 import static org.kie.remote.CommonConfig.LOCAL_MESSAGE_SYSTEM_CONF;
 import static org.kie.remote.util.ConfigurationUtil.readBoolean;
 
-/*Duplicated until moved in the ha-core-infra*/
 public class InfraFactory {
 
-    private InfraFactory() { }
-
-    public static EventConsumer getEventConsumer(EnvConfig config) {
-        return config.isLocal() ? new LocalConsumer(config) : new DefaultKafkaConsumer(config, getProducer(false));
-    }
-
-    public static SessionSnapshooter getSnapshooter(EnvConfig envConfig) {
-        return new DefaultSessionSnapShooter(envConfig);
-    }
-
-    public static ConsumerHandler getConsumerHandler(Producer producer, EnvConfig envConfig) {
-        return new DroolsConsumerHandler(producer, envConfig, getSnapshooter(envConfig), new ConsumerUtilsCoreImpl());
-    }
+    private InfraFactory(){}
 
     public static KafkaConsumer getConsumer(String topic, Properties properties) {
         KafkaConsumer consumer = new KafkaConsumer(properties);
@@ -93,7 +69,17 @@ public class InfraFactory {
         return consumer;
     }
 
+    public static Listener getListener(Properties props, boolean isLocal){
+        return new Listener(props, InfraFactory.getListenerThread(TopicsConfig.getDefaultTopicsConfig(), isLocal, props));
+    }
 
+    public static ListenerThread getListenerThread(TopicsConfig topicsConfig,
+                                                   boolean isLocal,
+                                                   Properties configuration) {
+        return isLocal ?
+                new LocalListenerThread(topicsConfig) :
+                new KafkaListenerThread(getMergedConf(configuration), topicsConfig);
+    }
 
     public static Properties getMergedConf(Properties configuration) {
         Properties conf = ClientUtils.getConfiguration(ClientUtils.CONSUMER_CONF);
@@ -101,16 +87,12 @@ public class InfraFactory {
         return conf;
     }
 
-    public static RemoteKieSession createRemoteKieSession(Properties configuration,
-                                                          Listener listener,
-                                                          Producer producer) {
+
+    public static RemoteKieSession createRemoteKieSession(Properties configuration, Listener listener, Producer producer) {
         return new RemoteKieSessionImpl(configuration, listener, producer);
     }
 
-    public static RemoteKieSession createRemoteKieSession(Properties configuration,
-                                                          TopicsConfig envConfig,
-                                                          Listener listener,
-                                                          Producer producer) {
+    public static RemoteKieSession createRemoteKieSession(Properties configuration, TopicsConfig envConfig, Listener listener, Producer producer) {
         return new RemoteKieSessionImpl(configuration, envConfig, listener, producer);
     }
 
@@ -122,20 +104,12 @@ public class InfraFactory {
         return isLocal ? new LocalProducer() : new EventProducer();
     }
 
-    public static RemoteStreamingKieSession createRemoteStreamingKieSession(Properties configuration,
-                                                                            Listener listener,
-                                                                            Producer producer) {
+    public static RemoteStreamingKieSession createRemoteStreamingKieSession(Properties configuration, Listener listener, Producer producer) {
         return new RemoteStreamingKieSessionImpl(configuration, listener, producer);
     }
 
-    public static RemoteStreamingKieSession createRemoteStreamingKieSession(Properties configuration,
-                                                                            TopicsConfig envConfig,
-                                                                            Listener listener,
-                                                                            Producer producer) {
+    public static RemoteStreamingKieSession createRemoteStreamingKieSession(Properties configuration, TopicsConfig envConfig, Listener listener, Producer producer) {
         return new RemoteStreamingKieSessionImpl(configuration, envConfig, listener, producer);
     }
 
-    public static ItemToProcess getItemToProcess(ConsumerRecord record) {
-        return new ItemToProcess(record.key().toString(), record.offset(), record.value());
-    }
 }
