+++ hacep/pom.xml
@@ -13,11 +13,6 @@
   <dependencies>
     <dependency>
       <groupId>org.kie</groupId>
-      <artifactId>ha-core-infra</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.kie</groupId>
       <artifactId>kie-remote</artifactId>
     </dependency>
