+++ hacep/src/test/java/org/kie/hacep/exceptions/ExceptionsTest.java
