+++ integration/src/test/java/org/kie/hacep/core/InfraFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.remote;
+package org.kie.hacep.core;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,7 +24,20 @@ import java.util.Set;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.common.PartitionInfo;
 import org.apache.kafka.common.TopicPartition;
-
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.kie.hacep.EnvConfig;
+import org.kie.hacep.consumer.DroolsConsumerHandler;
+import org.kie.hacep.core.infra.DefaultSessionSnapShooter;
+import org.kie.hacep.core.infra.SessionSnapshooter;
+import org.kie.hacep.core.infra.consumer.ConsumerHandler;
+import org.kie.hacep.core.infra.consumer.DefaultKafkaConsumer;
+import org.kie.hacep.core.infra.consumer.EventConsumer;
+import org.kie.hacep.core.infra.consumer.ItemToProcess;
+import org.kie.hacep.core.infra.consumer.LocalConsumer;
+import org.kie.hacep.util.ConsumerUtilsCoreImpl;
+import org.kie.remote.RemoteKieSession;
+import org.kie.remote.RemoteStreamingKieSession;
+import org.kie.remote.TopicsConfig;
 import org.kie.remote.impl.ClientUtils;
 import org.kie.remote.impl.RemoteKieSessionImpl;
 import org.kie.remote.impl.RemoteStreamingKieSessionImpl;
@@ -38,11 +51,23 @@ import org.kie.remote.impl.producer.Producer;
 
 import static org.kie.remote.CommonConfig.LOCAL_MESSAGE_SYSTEM_CONF;
 import static org.kie.remote.util.ConfigurationUtil.readBoolean;
-
-/**Duplicated until the mddules are moved*/
+/*Duplicated until moved into ha-core-infra*/
 public class InfraFactory {
 
     private InfraFactory(){}
+
+    public static EventConsumer getEventConsumer(EnvConfig config) {
+        return config.isLocal() ? new LocalConsumer(config) : new DefaultKafkaConsumer(config, getProducer(false));
+    }
+
+    public static SessionSnapshooter getSnapshooter(EnvConfig envConfig) {
+        return new DefaultSessionSnapShooter(envConfig);
+    }
+
+    public static ConsumerHandler getConsumerHandler(Producer producer, EnvConfig envConfig) {
+        return new DroolsConsumerHandler(producer, envConfig, getSnapshooter(envConfig), new ConsumerUtilsCoreImpl());
+    }
+
 
     public static KafkaConsumer getConsumer(String topic, Properties properties) {
         KafkaConsumer consumer = new KafkaConsumer(properties);
@@ -111,6 +136,10 @@ public class InfraFactory {
 
     public static RemoteStreamingKieSession createRemoteStreamingKieSession(Properties configuration, TopicsConfig envConfig, Listener listener, Producer producer) {
         return new RemoteStreamingKieSessionImpl(configuration, envConfig, listener, producer);
+    }
+
+    public static ItemToProcess getItemToProcess(ConsumerRecord record) {
+        return new ItemToProcess(record.key().toString(), record.offset(), record.value());
     }
 
 }
