+++ kie-remote/pom.xml
@@ -13,10 +13,6 @@
 
   <dependencies>
       <dependency>
-          <groupId>org.kie</groupId>
-          <artifactId>ha-core-infra</artifactId>
-      </dependency>
-      <dependency>
           <groupId>org.apache.kafka</groupId>
           <artifactId>kafka-clients</artifactId>
       </dependency>
