+++ sample-hacep-project/sample-hacep-project-client/pom.xml
@@ -14,6 +14,10 @@
     <dependencies>
         <dependency>
             <groupId>org.kie</groupId>
+            <artifactId>ha-core-infra</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.kie</groupId>
             <artifactId>kie-remote</artifactId>
         </dependency>
         <dependency>
